refactor: pull out caclulations to util which we can use for all aspects
